fix(codeql): break runner↔checkpoint cyclic import via models.py
Move Breakpoint and CallFrame dataclasses to a new models.py module.
runner.py and checkpoint.py both import from models.py, eliminating
the mutual dependency that triggered py/unsafe-cyclic-import alerts.
