+++ packages/core/src/rpaforge/core/checkpoint.py
@@ -14,10 +14,9 @@ from dataclasses import dataclass, field
 from datetime import datetime
 from enum import Enum
 from pathlib import Path
-from typing import TYPE_CHECKING, Any
+from typing import Any
 
-if TYPE_CHECKING:
-    from rpaforge.core.runner import Breakpoint, CallFrame
+from rpaforge.core.models import Breakpoint, CallFrame
 
 logger = logging.getLogger("rpaforge")
