+++ packages/core/src/rpaforge/core/runner.py
@@ -9,7 +9,6 @@ from __future__ import annotations
 import logging
 import threading
 from collections.abc import Callable
-from dataclasses import dataclass
 from enum import Enum
 from typing import TYPE_CHECKING, Any
 
@@ -20,6 +19,7 @@ from rpaforge.core.execution import (
 )
 from rpaforge.core.executor import ProcessExecutor, StopExecution
 from rpaforge.core.interfaces import Executor
+from rpaforge.core.models import Breakpoint, CallFrame
 from rpaforge.core.safe_evaluator import safe_eval
 from rpaforge.core.validator import (
     ValidationError as ProcessValidationError,
@@ -40,30 +40,6 @@ class RunnerState(Enum):
     PAUSED = "paused"
     CANCELLING = "cancelling"
     STOPPED = "stopped"
-
-
-@dataclass
-class Breakpoint:
-    """A breakpoint in execution."""
-
-    id: str
-    node_id: str
-    line: int = 0
-    enabled: bool = True
-    condition: str | None = None
-    hit_count: int = 0
-    hit_condition: str | None = None
-
-
-@dataclass
-class CallFrame:
-    """Call stack frame."""
-
-    activity: str
-    library: str
-    line: int
-    args: tuple[Any, ...]
-    node_id: str = ""
 
 
 class ProcessRunner:
